edited zsh some more

+++ install/create_links.sh
@@ -9,13 +9,21 @@ echo ".dotfiles not found, please make sure this repo has been cloned in your ho
 exit 
 fi 
 
-echo "removing config files if they already exist"
-# todo maybe have a prompt or something
-rm -rf $HOME/.config
-rm -rf $HOME/.vimrc
-rm -rf $HOME/.gitconfig
-rm -rf $HOME/.zshrc
-rm -rf $HOME/.zshenv
+echo "this script is going to overwrite all your current config files. you cool with that? (y/n)"
+read overwrite
+
+if [ $overwrite = "y" ] 
+then
+	echo "alright then, deleting"
+	rm -rf $HOME/.config
+	rm -rf $HOME/.vimrc
+	rm -rf $HOME/.gitconfig
+	rm -rf $HOME/.zshrc
+	rm -rf $HOME/.zshenv
+else
+	echo "alright, goodbye then"
+	exit
+fi
 
 mkdir -p $HOME/.config
 ln -s $DOTFILES/vimrc $HOME/.vimrc
@@ -28,6 +36,8 @@ ln -s $DOT_CONFIG/neofetch $HOME_CONFIG/neofetch
 echo "symlinked neofetch in .config"	
 ln -s $DOT_CONFIG/nvim $HOME_CONFIG/nvim
 echo "symlinked nvim in .config"	
+# vscode
+# firefox
 
 echo "symlinked everything"
